Refactor date and time utilities in date-utils.ts and move functions to dates.ts
- Consolidated date parsing and formatting logic by moving `parseTimeString`, `formatDateForStorage`, and `formatTimeForStorage` from `date-utils.ts` to `dates.ts`, enhancing modularity.
- Updated `formatDateForDisplay` to utilize a new Temporal-based helper for improved date parsing and error handling.
- Refactored `parseDateString` to ensure robust logging and fallback mechanisms for date reconstruction failures.

These changes improve the reliability and clarity of date and time handling in the Expo app.
